Changes to Portfolio/Resume
I uploaded an updated resume, as well as updated the website with some recent Medium Articles. I'm going to update this again after I've written the other articles.

Note to self: When I reach 10+ articles on Medium, my website needs a 'Writing' section of my portfolio (as in, I need another option in the navbar).